pyproject.toml
[tool.poetry]
name = "djanet"
version = "0.1.0"
description = "Social network by Django"
authors = ["Sergey_Zaitsew"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
djangorestframework = "^3.11.1"
pillow = "^7.2.0"
psycopg2 = "^2.8.5"
djoser = "^2.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
